chore #19: Domain - Entity로 이동
