+++ dogether/Domain/Entity/ReviewResults.swift
@@ -1,0 +1,13 @@
+//
+//  ReviewResults.swift
+//  dogether
+//
+//  Created by seungyooooong on 2/12/25.
+//
+
+import Foundation
+
+enum ReviewResults: String {
+    case approve = "APPROVE"
+    case reject = "REJECT"
+}
